lovable-ocr/requirements.txt
fastapi
uvicorn
pydantic
albumentations
transformers
torch
Pillow
numpy
pix2tex
python-multipart
